bus-ticket-app/microservices/user-auth/requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
PyJWT==2.7.0
